// src/api/shared/common/screen-code.enum.ts
export enum ScreenCode {
  ORDERS_VIEW = "orders.view",
  ORDERS_LIST = "orders.list",
  ORDERS_CREATE = "orders.create",
  ORDERS_SHOW = "orders.show",
  ORDERS_EDIT = "orders.edit",
  POSITIONS_LIST = "positions.list",
  POSITIONS_CREATE = "positions.create",
  POSITIONS_EDIT = "positions.edit",
  POSITIONS_RESERVE = "positions.reserve",
}